(refactor) use method sources from external class in util
