+++ link-parser/src/test/java/ru/tinkoff/edu/java/test/parser/GithubParserTest.java
@@ -1,24 +1,21 @@
 package ru.tinkoff.edu.java.test.parser;
 
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 import ru.tinkoff.edu.java.parser.GitHubParser;
 import ru.tinkoff.edu.java.parser.response.BaseResponse;
 import ru.tinkoff.edu.java.parser.response.GitHubResponse;
 
-import static org.assertj.core.api.Assertions.*;
-import static org.junit.jupiter.api.Assertions.*;
-
 import java.util.Optional;
-import java.util.stream.Stream;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.assertAll;
 
 public class GithubParserTest {
     GitHubParser gitHubParser = new GitHubParser();
 
     @ParameterizedTest
-    @MethodSource("provideValidLinks")
+    @MethodSource("ru.tinkoff.edu.java.test.parser.util.TestSamples#provideGitHubValidLinks")
     void parse_returnGitHubResponse_ValidURLs(String url, String expectedUser, String expectedRepo) {
         // given
 
@@ -36,7 +33,7 @@ public class GithubParserTest {
     }
 
     @ParameterizedTest
-    @MethodSource("provideInvalidLinks")
+    @MethodSource("ru.tinkoff.edu.java.test.parser.util.TestSamples#provideInvalidLinks")
     void parse_returnEmptyOptional_InvalidURLs(String url) {
         // given
 
@@ -46,22 +43,4 @@ public class GithubParserTest {
         // then
         assertThat(response).isEmpty();
     }
-
-    static Stream<Arguments> provideValidLinks() {
-        return TestSamples.provideGitHubValidLinks();
-    }
-
-    static Stream<String> provideInvalidLinks() {
-        return Stream.of(
-                "http://github.com/emil/project   ",
-                "https://github.co/emil/project",
-                "github.com/user/repo",
-                "https://github.com//",
-                "https://github.com/ /repo",
-                "https://github.com/u s e r/repo",
-                "https://vk.com/user/repo"
-        );
-    }
-
-
 }
