+++ link-parser/src/test/java/ru/tinkoff/edu/java/test/parser/ParserChainTest.java
@@ -1,7 +1,6 @@
 package ru.tinkoff.edu.java.test.parser;
 
 import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 import ru.tinkoff.edu.java.parser.GitHubParser;
 import ru.tinkoff.edu.java.parser.LinkChainParser;
@@ -12,7 +11,6 @@ import ru.tinkoff.edu.java.parser.response.GitHubResponse;
 import ru.tinkoff.edu.java.parser.response.StackOverflowResponse;
 
 import java.util.Optional;
-import java.util.stream.Stream;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
@@ -21,9 +19,8 @@ public class ParserChainTest {
 
     LinkParser linkParser = LinkChainParser.chain(new GitHubParser(), new StackOverflowParser());
 
-
     @ParameterizedTest
-    @MethodSource("provideGitHubValidLinks")
+    @MethodSource("ru.tinkoff.edu.java.test.parser.util.TestSamples#provideGitHubValidLinks")
     void parse_returnGitHubResponse_ValidURLs(String url, String expectedUser, String expectedRepo) {
         // given
 
@@ -42,7 +39,7 @@ public class ParserChainTest {
 
 
     @ParameterizedTest
-    @MethodSource("provideStackOverflowValidLinks")
+    @MethodSource("ru.tinkoff.edu.java.test.parser.util.TestSamples#provideStackOverflowValidLinks")
     void parse_returnStackOverflowResponse_ValidURLs(String url, String expectedQuestionId) {
         // given
 
@@ -60,7 +57,7 @@ public class ParserChainTest {
     }
 
     @ParameterizedTest
-    @MethodSource("provideInvalidLinks")
+    @MethodSource("ru.tinkoff.edu.java.test.parser.util.TestSamples#provideInvalidLinks")
     void parse_returnEmptyOptional_InvalidURLs(String url) {
         // given
 
@@ -70,32 +67,4 @@ public class ParserChainTest {
         // then
         assertThat(response).isEmpty();
     }
-
-
-    static Stream<Arguments> provideGitHubValidLinks() {
-        return TestSamples.provideGitHubValidLinks();
-    }
-
-    static Stream<Arguments> provideStackOverflowValidLinks() {
-        return TestSamples.provideStackOverflowValidLinks();
-    }
-
-    static Stream<String> provideInvalidLinks() {
-        return Stream.of(
-                "http://github.com/emil/project",
-                "https://stackoverflow.com/questions/",
-                "https://stackoverflow.com/questions/      ",
-                "https://stackoverflow.com/questions/aaaa",
-                "https://stackoverflow.com/questions/123 123",
-                "https://vk.com/user/repo",
-                "http://github.com/emil/project   ",
-                "https://github.co/emil/project",
-                "github.com/user/repo",
-                "https://github.com//",
-                "https://github.com/ /repo",
-                "https://github.com/u s e r/repo",
-                "https://vk.com/user/repo"
-        );
-    }
-
 }
