+++ link-parser/src/test/java/ru/tinkoff/edu/java/test/parser/util/TestSamples.java
@@ -1,4 +1,4 @@
-package ru.tinkoff.edu.java.test.parser;
+package ru.tinkoff.edu.java.test.parser.util;
 
 import org.junit.jupiter.params.provider.Arguments;
 import ru.tinkoff.edu.java.parser.response.GitHubResponse;
@@ -12,7 +12,7 @@ public class TestSamples {
     private final static String URL = "https://stackoverflow.com/questions/%s";
     private final static String SLASH_URL = URL + "/";
 
-    static Stream<Arguments> provideGitHubValidLinks() {
+    public static Stream<Arguments> provideGitHubValidLinks() {
 
         Stream<GitHubResponse> input = Stream.of(
                 new GitHubResponse("user", "repo"),
@@ -29,13 +29,30 @@ public class TestSamples {
         );
     }
 
-    static Stream<Arguments> provideStackOverflowValidLinks() {
-
+    public static Stream<Arguments> provideStackOverflowValidLinks() {
         Stream<String> questions = Stream.of("12356689", "945987459", "0");
 
         return questions.flatMap(question -> Stream.of(
                 Arguments.of(URL.formatted(question), question),
                 Arguments.of(SLASH_URL.formatted(question), question)));
+    }
+
+    public static Stream<String> provideInvalidLinks() {
+        return Stream.of(
+                "http://github.com/emil/project",
+                "https://stackoverflow.com/questions/",
+                "https://stackoverflow.com/questions/      ",
+                "https://stackoverflow.com/questions/aaaa",
+                "https://stackoverflow.com/questions/123 123",
+                "https://vk.com/user/repo",
+                "http://github.com/emil/project   ",
+                "https://github.co/emil/project",
+                "github.com/user/repo",
+                "https://github.com//",
+                "https://github.com/ /repo",
+                "https://github.com/u s e r/repo",
+                "https://vk.com/user/repo"
+        );
     }
 
 }
